refactor: 가독성을 높이기 위한 파일명 변경 event -> controller
