+++ src/controller/initEvent.js
