+++ src/controller/line/removeLineEvent.js
