+++ src/controller/section/addSectionEvent.js
