+++ src/controller/section/selectLineButtonEvent.js
