+++ src/controller/station/removeStationEvent.js
